+++ src/pykeen/models/baseline/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+"""Non-parametric baseline models."""
+
+from .models import EvaluationOnlyModel, MarginalDistributionBaseline, SoftInverseTripleBaseline  # noqa:F401
