+++ src/pykeen/models/baseline/models.py
@@ -1,0 +1,184 @@
+# -*- coding: utf-8 -*-
+
+"""Non-parametric baseline models."""
+
+from abc import ABC
+from typing import Optional
+
+import numpy
+import scipy.sparse
+import torch
+
+from .utils import get_relation_similarity, marginal_score, get_csr_matrix
+from ..base import Model
+from ...triples import CoreTriplesFactory
+
+__all__ = [
+    'EvaluationOnlyModel',
+    'MarginalDistributionBaseline',
+    'SoftInverseTripleBaseline',
+]
+
+
+class EvaluationOnlyModel(Model, ABC):
+    """A model which only implements the methods used for evaluation."""
+
+    def __init__(self, triples_factory: CoreTriplesFactory):
+        """Non-parametric models take a minimal set of arguments.
+
+        :param triples_factory: The training triples factory is used to assign the number of entities, relations,
+            and inverse condition in the non-parametric model.
+        """
+        super().__init__(
+            triples_factory=triples_factory,
+            # These operations are deterministic and a random seed can be fixed
+            # just to avoid warnings
+            random_seed=0,
+            # These operations do not need to be performed on a GPU
+            preferred_device='cpu',
+        )
+
+    def _reset_parameters_(self):
+        """Non-parametric models do not implement :meth:`Model._reset_parameters_`."""
+        raise RuntimeError
+
+    def collect_regularization_term(self):  # noqa:D102
+        """Non-parametric models do not implement :meth:`Model.collect_regularization_term`."""
+        raise RuntimeError
+
+    def score_hrt(self, hrt_batch: torch.LongTensor):  # noqa:D102
+        """Non-parametric models do not implement :meth:`Model.score_hrt`."""
+        raise RuntimeError
+
+    def score_r(self, ht_batch: torch.LongTensor):  # noqa:D102
+        """Non-parametric models do not implement :meth:`Model.score_r`."""
+        raise RuntimeError
+
+
+class MarginalDistributionBaseline(EvaluationOnlyModel):
+    r"""
+    Score based on marginal distributions.
+
+    To predict scores for the tails, we simplify
+
+    .. math ::
+        P(t | h, r) = P(t | h) * P(t | r)
+
+    Depending on the settings, we either set $P(t | *) = \frac{1}{n}$, or estimate them by counting occurrences in the
+    training triples.
+
+    .. note ::
+        This model cannot make use of GPU acceleration, since internally it uses scipy's sparse matrices.
+
+    ---
+    name: Marginal Distribution Baseline
+    citation:
+        author: Berrendorf
+        year: 2021
+        link: https://github.com/pykeen/pykeen/pull/514
+        github: pykeen/pykeen
+    """
+
+    def __init__(
+        self,
+        triples_factory: CoreTriplesFactory,
+        entity_margin: bool = True,
+        relation_margin: bool = True,
+    ):
+        """
+        Initialize the model.
+
+        :param triples_factory:
+            The triples factory containing the training triples.
+        """
+        super().__init__(triples_factory=triples_factory)
+        h, r, t = numpy.asarray(triples_factory.mapped_triples).T
+        if relation_margin:
+            self.head_per_relation, self.tail_per_relation = [
+                get_csr_matrix(
+                    row_indices=r,
+                    col_indices=col_indices,
+                    shape=(triples_factory.num_relations, triples_factory.num_entities),
+                )
+                for col_indices in (h, t)
+            ]
+        else:
+            self.head_per_relation = self.tail_per_relation = None
+        if entity_margin:
+            self.head_per_tail, self.tail_per_head = [
+                get_csr_matrix(
+                    row_indices=row_indices,
+                    col_indices=col_indices,
+                    shape=(triples_factory.num_entities, triples_factory.num_entities),
+                )
+                for row_indices, col_indices in ((t, h), (h, t))
+            ]
+        else:
+            self.head_per_tail = self.tail_per_head = None
+
+    def score_t(self, hr_batch: torch.LongTensor) -> torch.FloatTensor:  # noqa:D102
+        return marginal_score(
+            entity_relation_batch=hr_batch,
+            per_entity=self.tail_per_head,
+            per_relation=self.tail_per_relation,
+            num_entities=self.num_entities,
+        )
+
+    def score_h(self, rt_batch: torch.LongTensor) -> torch.FloatTensor:  # noqa:D102
+        return marginal_score(
+            entity_relation_batch=rt_batch.flip(1),
+            per_entity=self.head_per_tail,
+            per_relation=self.head_per_relation,
+            num_entities=self.num_entities,
+        )
+
+
+class SoftInverseTripleBaseline(EvaluationOnlyModel):
+    """Score based on relation similarity.
+
+    ---
+    name: Soft Inverse Triple Baseline
+    citation:
+        author: Berrendorf
+        year: 2021
+        link: https://github.com/pykeen/pykeen/pull/514
+        github: pykeen/pykeen
+    """
+
+    def __init__(
+        self,
+        triples_factory: CoreTriplesFactory,
+        threshold: Optional[float] = None,
+    ):
+        super().__init__(triples_factory=triples_factory)
+        # compute relation similarity matrix
+        self.sim = get_relation_similarity(triples_factory, to_inverse=False, threshold=threshold)
+        self.sim_inv = get_relation_similarity(triples_factory, to_inverse=True, threshold=threshold)
+
+        mapped_triples = numpy.asarray(triples_factory.mapped_triples)
+        self.rel_to_head = scipy.sparse.coo_matrix(
+            (
+                numpy.ones(shape=(triples_factory.num_triples,), dtype=numpy.float32),
+                (mapped_triples[:, 1], mapped_triples[:, 0]),
+            ),
+            shape=(triples_factory.num_relations, triples_factory.num_entities),
+        ).tocsr()
+        self.rel_to_tail = scipy.sparse.coo_matrix(
+            (
+                numpy.ones(shape=(triples_factory.num_triples,), dtype=numpy.float32),
+                (mapped_triples[:, 1], mapped_triples[:, 2]),
+            ),
+            shape=(triples_factory.num_relations, triples_factory.num_entities),
+        ).tocsr()
+
+    def score_t(self, hr_batch: torch.LongTensor) -> torch.FloatTensor:  # noqa:D102
+        r = hr_batch[:, 1]
+        scores = self.sim[r, :] @ self.rel_to_tail + self.sim_inv[r, :] @ self.rel_to_head
+        scores = numpy.asarray(scores.todense())
+        return torch.from_numpy(scores)
+
+    def score_h(self, rt_batch: torch.LongTensor) -> torch.FloatTensor:  # noqa:D102
+        r = rt_batch[:, 0]
+        scores = self.sim[r, :] @ self.rel_to_head + self.sim_inv[r, :] @ self.rel_to_tail
+        scores = numpy.asarray(scores.todense())
+        return torch.from_numpy(scores)
